# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "URL Shortener API Documentation",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/": {
            "get": {
                "summary": "Can be used to know if the server is running.",
                "operationId": "health-check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Message"
                        }
                    }
                }
            }
        },
        "/url": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "summary": "receives a url and return its short representation.",
                "operationId": "long-to-short",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/types.Url"
                        }
                    }
                }
            }
        },
        "/url/{url}": {
            "get": {
                "summary": "receives a short url as uri parameter and return w/ a redirect to the long url previously registered.",
                "operationId": "short-to-long",
                "parameters": [
                    {
                        "type": "string",
                        "description": "url",
                        "name": "url",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "307": {
                        "description": "Temporary Redirect",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "types.Message": {
            "type": "object",
            "properties": {
                "msg": {
                    "type": "string"
                }
            }
        },
        "types.Url": {
            "type": "object",
            "properties": {
                "url": {
                    "type": "string"
                }
            }
        }
    }
}